refactoring show decription or clue
